feat(schema): move section grouping helpers to separate file
